Fix alert borrar reviews, fix toLowerCase en postAllBooks

+++ client/src/components/Review/Review.jsx
@@ -91,13 +91,26 @@ export default function Review({ id }) {
 
   function handleDelete(e) {
     e.preventDefault();
-    dispatch(deleteReviews({ UserId: user.uid, BookId: id }));
+    Swal.fire({
+      title: '¡Cuidado! Estás a punto de borrar tu reseña.',
+      width: 650,
+      text: "¿Quieres borrar tu reseña? Esto no podrá deshacerse.",
+      icon: 'warning',
+      iconColor: "#355070",
+      showCancelButton: true,
+      background: "#19191a",
+      color: "#e1e1e1",
+      confirmButtonColor: '#355070',
+      cancelButtonColor: '#B270A2',
+      confirmButtonText: '¡Si! Borrar reseña',
+      cancelButtonText: "Cancelar"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteReviews({ UserId: user.uid, BookId: id }));
 
-    setButtonDisabled(true);
-    setTimeout(function () {
-      setButtonDisabled(false);
-    }, 1000);
-    swalAlert(3000, "success", "Has eliminado tu reseña, puedes agregar una nueva!")
+        swalAlert(3000, "success", "Has eliminado tu reseña, puedes agregar una nueva!")
+      }
+    })
   }
 
   useEffect(() => {
